# microservices/aspnetcore/team/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Allwissend Team API",
    "version" : "1.0.0",
    "title" : "Allwissend Microservices API"
  },
  "host" : "backend.application.dev",
  "basePath" : "/v1",
  "schemes" : [ "http", "https" ],
  "produces" : [ "application/json" ],
  "security" : [ {
    "internalApiKey" : [ ]
  } ],
  "paths" : {
...
    "/team/{teamUuid}/member" : {
      "post" : {
        "tags" : [ "PractitionerCommand" ],
        "summary" : "Add a new Member to a Team",
        "description" : "The endpoint save a new *Team*.\n",
        "operationId" : "addTeamMember",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamUuid",
          "in" : "path",
          "description" : "Unique identifier representing a specific team that needs to be fetched",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Team to be saved",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Member"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbiden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "504" : {
            "description" : "Gateway Timeout",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/team/{teamUuid}/member/{practitionerUuid}" : {
      "put" : {
        "tags" : [ "PractitionerCommand" ],
        "summary" : "Updates a Member in a Team type",
        "description" : "The endpoint returns a list of *Organization*.\n",
        "operationId" : "updateTeamMember",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamUuid",
          "in" : "path",
          "description" : "Unique identifier representing a specific team that needs to be fetched",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "practitionerUuid",
          "in" : "path",
          "description" : "Unique identifier representing a specific practitioner that needs to be fetched",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Team to be saved",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Member"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Accepted"
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbiden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Resources Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Invalid input",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "504" : {
            "description" : "Gateway Timeout",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "PractitionerCommand" ],
        "summary" : "Delete a Member from a Team type",
        "description" : "The endpoint returns a list of *Organization*.\n",
        "operationId" : "deleteTeamMember",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "teamUuid",
          "in" : "path",
          "description" : "Unique identifier representing a specific team that needs to be fetched",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "name" : "practitionerUuid",
          "in" : "path",
          "description" : "Unique identifier representing a specific practitioner that needs to be fetched",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "202" : {
            "description" : "Accepted"
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbiden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Resources Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "504" : {
            "description" : "Gateway Timeout",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "default" : {
            "description" : "Unexpected error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "internalApiKey" : {
      "type" : "apiKey",
      "name" : "token",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Member" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Unique identifier representing a specific member"
        },
        "period" : {
          "$ref" : "#/definitions/Period"
        },
        "practitioner" : {
          "$ref" : "#/definitions/Practitioner"
        },
        "type" : {
          "type" : "string",
          "description" : "type....",
          "enum" : [ "permanent", "temporary", "loaned", "outsourced", "rented" ]
        },
        "status" : {
          "type" : "string",
          "description" : "Status...",
          "enum" : [ "evaluation", "not_started", "in_hiring", "in_training", "absent", "lunch", "present", "loaned", "working", "stopped", "loss", "injured", "away" ]
        }
      },
      "description" : "Member?!",
      "example" : {
        "period" : {
          "start" : "start",
          "end" : "end"
        },
        "practitioner" : {
          "name" : {
            "given" : [ "given", "given" ],
            "lastUpdated" : "lastUpdated",
            "prefix" : [ "prefix", "prefix" ],
            "alias" : "alias",
            "family" : "family",
            "suffix" : [ "suffix", "suffix" ],
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "first" : "first",
            "creation" : "creation"
          },
          "active" : true,
          "photo" : [ {
            "lastUpdated" : "lastUpdated",
            "extension" : "extension",
            "data" : "data",
            "size" : 0,
            "language" : "language",
            "title" : "title",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "contentType" : "contentType",
            "url" : "url",
            "hash" : "hash",
            "creation" : "creation"
          }, {
            "lastUpdated" : "lastUpdated",
            "extension" : "extension",
            "data" : "data",
            "size" : 0,
            "language" : "language",
            "title" : "title",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "contentType" : "contentType",
            "url" : "url",
            "hash" : "hash",
            "creation" : "creation"
          } ],
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "type" : "permanent",
        "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "status" : "evaluation"
      }
    },
    "Practitioner" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Unique identifier representing a specific practitioner."
        },
        "active" : {
          "type" : "boolean",
          "description" : "To do"
        },
        "name" : {
          "$ref" : "#/definitions/Name"
        },
        "photo" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Attachment"
          },
          "minItems" : 1
        }
      },
      "example" : {
        "name" : {
          "given" : [ "given", "given" ],
          "lastUpdated" : "lastUpdated",
          "prefix" : [ "prefix", "prefix" ],
          "alias" : "alias",
          "family" : "family",
          "suffix" : [ "suffix", "suffix" ],
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "first" : "first",
          "creation" : "creation"
        },
        "active" : true,
        "photo" : [ {
          "lastUpdated" : "lastUpdated",
          "extension" : "extension",
          "data" : "data",
          "size" : 0,
          "language" : "language",
          "title" : "title",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "contentType" : "contentType",
          "url" : "url",
          "hash" : "hash",
          "creation" : "creation"
        }, {
          "lastUpdated" : "lastUpdated",
          "extension" : "extension",
          "data" : "data",
          "size" : 0,
          "language" : "language",
          "title" : "title",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "contentType" : "contentType",
          "url" : "url",
          "hash" : "hash",
          "creation" : "creation"
        } ],
        "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      }
    },
    "Attachment" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Unique identifier"
        },
        "extension" : {
          "type" : "string",
          "description" : "Additional Content defined by implementations"
        },
        "contentType" : {
          "type" : "string",
          "description" : "Mime type of the content, with charset etc."
        },
        "language" : {
          "type" : "string",
          "description" : "Human language of the content (BCP-47)"
        },
        "data" : {
          "type" : "string",
          "format" : "base64Binary",
          "description" : "Data inline, base64ed"
        },
        "url" : {
          "type" : "string",
          "description" : "Uri where the data can be found"
        },
        "size" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Number of bytes of content (if url provided)",
          "minimum" : 0
        },
        "hash" : {
          "type" : "string",
          "format" : "base64Binary",
          "description" : "Hash of the data (sha-1, base64ed)"
        },
        "title" : {
          "type" : "string",
          "description" : "Label to display in place of the data"
        },
        "creation" : {
          "type" : "string",
          "format" : "datetime",
          "description" : "Date attachment was first created"
        },
        "lastUpdated" : {
          "type" : "string",
          "format" : "datetime",
          "description" : "Date attachment was last updated"
        }
      },
      "example" : {
        "lastUpdated" : "lastUpdated",
        "extension" : "extension",
        "data" : "data",
        "size" : 0,
        "language" : "language",
        "title" : "title",
        "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "contentType" : "contentType",
        "url" : "url",
        "hash" : "hash",
        "creation" : "creation"
      }
    },
    "Name" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Unique identifier"
        },
        "alias" : {
          "type" : "string"
        },
        "first" : {
          "type" : "string"
        },
        "family" : {
          "type" : "string"
        },
        "given" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "prefix" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "suffix" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "creation" : {
          "type" : "string",
          "format" : "datetime",
          "description" : "Date name was first created"
        },
        "lastUpdated" : {
          "type" : "string",
          "format" : "datetime",
          "description" : "Date name was last updated"
        }
      },
      "example" : {
        "given" : [ "given", "given" ],
        "lastUpdated" : "lastUpdated",
        "prefix" : [ "prefix", "prefix" ],
        "alias" : "alias",
        "family" : "family",
        "suffix" : [ "suffix", "suffix" ],
        "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
        "first" : "first",
        "creation" : "creation"
      }
    },
...
    "Period" : {
      "type" : "object",
      "properties" : {
        "start" : {
          "type" : "string",
          "format" : "datetime"
        },
        "end" : {
          "type" : "string",
          "format" : "datetime"
        }
      },
      "description" : "Time period when contact was/is in use. Allows contact to be placed in historical context.",
      "example" : {
        "start" : "start",
        "end" : "end"
      }
    },
...
    "Error" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "fields" : {
          "type" : "string"
        }
      }
    }
  },
...
}